royal-dental-lounge/requirements.txt
# Royal Dental Lounge ??Python Backend Requirements
flask==3.0.0
twilio==8.10.0
python-dotenv==1.0.0
pymongo==4.6.1
requests==2.31.0
gunicorn==21.2.0
